stair_step cython performance improvement/modified tests

+++ fast_intensity/fast_intensity.py
@@ -10,6 +10,25 @@ pyximport.install(setup_args={"include_dirs": np.get_include()})
 from .stair_step import stair_step
 from .fast_hist import fast_hist
 
+#
+# def stair_step(x, y, xp, yp):
+#     """
+#     Previous neighbor interpolation.
+#
+#     Args:
+#         x (list or np.array of numbers): sample points
+#         y (list or np.array of numbers): sample values (same size as x)
+#         xp (list or np.array of numbers): query points
+#         yp (list or np.array of numbers): preallocated list or np.array for
+#             query values (same size as xp)
+#
+#     Returns:
+#         np.array of interpolated values (float)
+#     """
+#     y = np.concatenate( ([0], y))
+#     return y[[np.where(x <= p)[0].max() for p in xp]]
+
+
 class FastIntensity(object):
     """Estimates (potentially nonstationary) event intensity vs. time.
 
@@ -66,7 +85,7 @@ class FastIntensity(object):
             end_event (number): end of the computed inference time range
         """
         # Convert to numpy array
-        events = np.array(events)
+        events = np.array(events, dtype=np.float)
         # Cut out of bounds events
         events = np.delete(events, np.where(events < start_event))
         events = np.delete(events, np.where(events > end_event))
@@ -183,7 +202,7 @@ class FastIntensity(object):
         self.grid = self._generate_grid(resolution, density)
 
         meanvals = np.zeros(len(self.grid))
-        vals = np.zeros(len(self.grid))
+        vals = np.zeros(len(self.grid), dtype=np.float)
         n = len(self.events) + 1
         num_edges = int(2 + np.max([2, np.ceil(density * n)]))
 
@@ -216,6 +235,6 @@ class FastIntensity(object):
 
         y[1:-1] = np.interp(w, np.linspace(0, n, n+1), events_w_endpoints)
 
-        y = np.array(y)
+        y = np.array(y, dtype=np.float)
         y = np.round(y/resolution)*resolution
         return y
